requirements.txt
mkdocs-material>=9.0
pytest
hypothesis
pyyaml
